pyproject.toml
[tool.poetry]
name = "nextdns-mcp"
version = "2.0.0"
description = "NextDNS MCP Server - Model Context Protocol server for NextDNS API built with FastMCP"
authors = ["NextDNS MCP Contributors"]
readme = "README.md"
packages = [{include = "nextdns_mcp", from = "src"}]
include = ["src/nextdns_mcp/*.yaml"]

[tool.poetry.dependencies]
python = "^3.13"
fastmcp = "^2.12.0"
httpx = "^0.28.1"
pyyaml = "^6.0"
python-dotenv = "^1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^1.2.0"
pytest-cov = "^4.1.0"
black = "^24.0.0"
ruff = "^0.3.0"
isort = "^7.0.0"
mypy = "^1.18.2"
types-pyyaml = "^6.0.12.20250915"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.black]
line-length = 100
target-version = ["py313"]

[tool.ruff]
line-length = 100
target-version = "py313"
